Moved mapper code to participant-mapper
