+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/individual-journey/services/individual-model-mapper.ts
@@ -2,6 +2,7 @@ import { SuitabilityAnswer, HasAccessToCamera, Hearing } from '../../base-journe
 import { IndividualSuitabilityModel } from '../individual-suitability.model';
 import { HearingSuitabilityResponse, HearingSuitabilityAnswer } from 'src/app/services/clients/api-client';
 import { MutableIndividualSuitabilityModel } from '../mutable-individual-suitability.model';
+import { ParticipantModelMapper } from '../../base-journey/services/participant-model-mapper';
 
 export const IndividualQuestionKeys = {
     AboutYou: 'ABOUT_YOU',
@@ -13,7 +14,8 @@ export const IndividualQuestionKeys = {
     Camera: 'CAMERA_MICROPHONE'
 };
 
-export class IndividualModelMapper {
+export class IndividualModelMapper extends ParticipantModelMapper {
+
     map(response: HearingSuitabilityResponse): IndividualSuitabilityModel {
         const model = new MutableIndividualSuitabilityModel();
         model.hearing = new Hearing(response.hearing_id, response.hearing_scheduled_at);
@@ -27,40 +29,5 @@ export class IndividualModelMapper {
         model.camera = this.mapComputerCamera(response.answers);
         return model;
     }
-    private mapBooleanValue(answers: HearingSuitabilityAnswer[], key: string) {
-        const answer = answers.find(a => a.question_key === key);
-        if (answer) {
-            return answer.answer === 'true';
-        }
-        return undefined;
-    }
-    private mapComputerCamera(answers: HearingSuitabilityAnswer[]): HasAccessToCamera {
-        const answer = answers.find(a => a.question_key === IndividualQuestionKeys.Camera);
-        if (answer) {
-            switch (answer.answer) {
-                case 'Yes':
-                    return HasAccessToCamera.Yes;
-                case 'No':
-                    return HasAccessToCamera.No;
-                case 'Not sure':
-                    return HasAccessToCamera.NotSure;
-                default:
-                    throw new Error(`Unexpected answer to computer question: ${answer.answer}`);
-            }
-        }
-        return undefined;
-    }
-    private mapBooleanAnswerFromKey(key: string, answers: HearingSuitabilityAnswer[]): SuitabilityAnswer {
-        const answer = answers.find(a => a.question_key === key);
-        if (answer) {
-            return this.answer(answer.answer, answer.extended_answer);
-        }
-        return new SuitabilityAnswer();
-    }
-    private answer(value: string, notes: string): SuitabilityAnswer {
-        const answer = new SuitabilityAnswer();
-        answer.answer = value === 'true';
-        answer.notes = notes;
-        return answer;
-    }
+
 }
